颇具雏形，能正常请求视频

+++ src/http/Database.h
@@ -10,7 +10,7 @@
 class Database {
 public:
     Database(std::string_view host, std::string_view user, std::string_view password, std::string_view database,
-             unsigned int port, std::string_view unixSocket, unsigned long clientFlag);
+             unsigned short port, std::string_view unixSocket, unsigned long clientFlag);
 
     Database(const Database &) = delete;
 
@@ -20,7 +20,7 @@ private:
     auto initialize(std::source_location sourceLocation = std::source_location::current()) -> void;
 
     auto connect(std::string_view host, std::string_view user, std::string_view password, std::string_view database,
-                 unsigned int port, std::string_view unixSocket, unsigned long clientFlag,
+                 unsigned short port, std::string_view unixSocket, unsigned long clientFlag,
                  std::source_location sourceLocation = std::source_location::current()) -> void;
 
 public:
